requirements.txt
streamlit==1.24.0
openai==0.27.2
pydantic==1.10.11
pyyaml==6.0.2
python-dotenv==1.0.1
anthropic==0.36.2